+++ handlers/background/__init__.py
@@ -1,9 +1,11 @@
 from .mute import check_mute
 from .last_modified_check import close_check
 from .delete_notify import notify_delete
+from .resolution_prompt import resolution_prompt_check
 
 __all__ = [
     'check_mute',
     'close_check',
-    'notify_delete'
+    'notify_delete',
+    'resolution_prompt_check',
 ]